+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v2.9 - 2022/08/16
+
+* Fix website field
+* Now the host,services and vulns create have a command associated
+
+# v2.8 - 2022/07/06
+
+* Add refresh workspace bottom
+
 # v2.7 - 2021/07/21
 
 * Use Faraday api version 3